video: implement hardware rendering with proper intermediate upscaling
We now have two textures, an intermediate 320x200 texture that we load
the RGBA surface into and another upscaled one. We render the former
into the latter using "nearest" integer upscaling and render the latter
to screen using "linear" scaling.

This way, the screen still looks gritty and pixelated but the numbers
in the status bar are properly interpolated. Should we ever want it even
grittier, we could increase the UPSCALE macro to e.g. 3, but I have
not found any any differences using higher numbers than that. Anyway,
rendering now out-performs current software scaling for any given
resolution on my system. \o/

Thanks to @EMue for the gold tip regarding the second texture and the
rendering quality settings applied to it. In fact, "the order matters".
;)

@fragglet Now look at this! ^^