input: Initial implementation of text input.
This should allow us to support eg. tablet devices with on-screen
keyboards in the future, but the implementation is kind of a hack.
Extend the ev_keydown event to support a third data type that
contains fully modified input characters. As a proof of concept,
switch over the Doom multiplayer chat code to use it.