doom: Correctly handle savegame fopen() failure.
If the savegame directory is unwriteable we should exit with an error -
the previous behavior caused an infinite loop of attempting to open the
same file repeatedly. Because the player may value the progress they
have made, attempt to write the savegame to a temporary directory first
before exiting with an error.

This fixes #659. Thanks to @terrorcide for the bug report.