textscreen: Add support for OS X high-DPI rendering.
When using a Mac with a retina display, we want to use the large
textscreen font, albeit in a normal-sized window. The HIGHDPI flag is
the way to do this properly with SDL2; try to set it and if we get the
output screen size we're hoping for, switch to the large font.

As part of this, bump the SDL2 dependency to 2.0.1, since the
SDL_WINDOW_ALLOW_HIGHDPI flag was only added in this version.