textscreen: Set linear scaling for texture buffer.
This is important for OS X - if multiple displays are connected, not
all of them may be retina displays, so the window may be rendered as
though it is always on a retina display. In this case, it's important
to use linear scaling as the default nearest scaling can cause the
large font to be distorted.