diskicon: Fix Strife disk icon.
Strife has no STCDROM lump; it always uses STDISK. Change the
V_EnableLoadingDisk() function to take the lump name as an argument,
as the lump to use for the loading disk is really game-specific.

Also fix the location where the Strife disk icon is shown on screen;
the vertical position wasn't quite right.