setup: Hide window sizes when fullscreen=true.
When the game is configured to run in full screen mode it is misleading
to keep showing the window size selectors as this can give the
impression that a fullscreen video mode is being selected. Instead,
hide the window sizes so that they only appear when running in windowed
mode. This fixes #717.