setup: Fix crash with Chex Quest warp select.
The level selection dialog was crashing on open when Chex Quest was
selected as an IWAD. This appears to have been caused by 4a829f22
which changed Chex Quest to be considered as a retail IWAD rather
than a shareware one, but did not update the tables in d_iwad.c to
match. Change the tables there to match and fix the crash.