video: Set pixel format correctly for rgbabuffer.
When drawing the screen we convert from the 8-bit surface to the
32-bit RGBA surface and then import the surface data into the texture
for hardware rendering. However, the RGBA mask for the 32-bit surface
must match the RGBA format for the screen otherwise the colors may
end up wrong.

Thanks to Ryan Freeman (CSonicGo) for reporting this problem on
Twitter: <https://twitter.com/Sneakernets/status/751612074967568388>
I have not tested this on a Raspberry Pi but I believe this should
fix the problem.