Add P_SubRandom and fix P_Random ambiguities
Doom/Heretic/Hexen frequently make use of a construct like

    P_Random() - P_Random()

The order of evaluation of which is undefined in C. Watcom always evaluated
left-to-right. Add a P_SubRandom routine inspired by the same solution
implemented in Eternity that encapsulates this order of evaluation and use that
in place of the above. Thanks @haleyjd

Clean up other occurrences of undefined evaluation order with P_Random,
including where the new P_SubRandom() would be used ambiguously (!), almost all
of which was in the Heretic and Hexen code. (@haleyjd had already cleaned up
Strife)

Fixes #404