SetVideoMode(): make re-entrant
Well, at least kind of (think of the forks that might want to change
renderer flags in-game). ;)

The SDL_Window *screen and the SDL_Surface *screenbuffer need not get
recreated, it's sufficient to create them once. The same applies to
the SDL_surface *rgbabuffer. Only the SDL_Renderer *renderer needs to
get recreated and the SDL_Texture *texture based on the new renderer.
The SDL_Texture *texture_upscaled needs to get recreated as well with
the new renderer. Else, it is recreated only if the texture scale
factors have changed.

All of this means no functional changes for Chocolate Doom, because it
only runs SetVideoMode() once at start up.