Add joystick virtual button for toggle automap
Add a distinct logic branch to the top of the AM_Responder
routines that exits out, rather than modify the rc variable/
branching logic in the rest of the routine. Use the same
button bounce timeout approach used for m_menu.

Additionally adjust the column widths for the setup screen for
joystick configuration, to account for the width of the string
"Toggle Automap".