video: Add video_display config option.
This new variable controls which SDL display the game window should be
shown on (currently untested as I don't have a second monitor to hand).
This only has an effect if in fullscreen mode or if window_position is
set to "center". As part of this, make that the new default for
window_position.

This is the first part of #734. See that bug for some context.