textscreen: Tweak TXT_GetKeyDescription() again.
We use the SDL APIs to get a localized name for keys now. However, we
must consider that SDL's key names are in UTF8 format and can therefore
contain non-ASCII characters which cannot be displayed on the CP437 text
screen. So double-check if key names contain any unprintable characters
and if they do, use a fallback set of default key names instead.

This fixes #840. Thanks to Julian Nechaevsky for the report.