win32: Auto-determine DLL path from LDFLAGS.
Copying DLLs raises the problem of finding where the DLLs are to copy.
With autotools, DLLs are usually installed to ${prefix}/bin alongside the
libraries installed to ${prefix}/lib. So we can use the -L arguments in
LDFLAGS passed to the linker to figure out a likely set of directories to
search in. Hook this into the build, too.

This fixes #764 - we now reliably build Windows .zip packages automatically
bundled with any DLLs which may be required.