video: Bump default pixel limit to 16M pixels.
This is enough to cover both the DCI 4K (4096 × 2160) and UHD-1
(3840 × 2160) 4K monitor standards. This should be okay since 16M * 4 =
64MB of texture RAM; it's rare (according to the Steam hardware survey)
to have less than 256MB of RAM nowadays.

This is a slight tweak to the fix on #859. It shouldn't make much, if
any, visual difference over 1600x1200, but if we can make use of the
extra RAM and gain a marginal improvement then I don't see any reason
not to.