video: move video_display check to I_GetWindowPosition()
This way make sure that the video_display variable is checked (1)
unconditionally and (2) early on in the code before other parts of the
video initialization path use it.

Formerly, the check was in CenterWindow() which was only called from
I_GetWindowPosition() if fullscreen was enabled or the window was
centered.

Fixes #898 thanks @jmtd