doom: Fix "Screen Size" label with BFG IWADs.
In the BFG Edition IWADs, M_SCRNSZ has been changed to contain the
text "Gamepad:" instead, which means that the options menu has a
screen size control that nonsensically has that label. Instead, swap
in the M_DISP lump when using the BFG Edition IWADs, so that it says
"Display" instead.