Add checks in P_RespawnSpecials().
When respawning a mobj we scan the mobjinfo[] array to find the type
that represents the object. It's possible that we don't find it and in
this case an array index overrun occurs. Handle this by exiting with an
error rather than continuing. This was detected by cppcheck static
analyses set up by @turol and blocking #939.