Add cppcheck static analysis to travis build (#939)
This runs cppcheck static analysis as part of the travis-ci checks. The
current Travis version of cppcheck doesn't appear to support all the checks
we want, but for now should be sufficient. Identified issues that would cause
these checks to fail have been resolved in previous commits.

Thanks to @turol for doing the work to set this up.