net: Add protocol negotiation on client connection.
The idea here is that we define a protocol versioning scheme, where we
give a symbolic (string) name to each protocol version. This allows us
to define new protocol versions in the future while retaining
compatibility between different Chocolate Doom versions. It also allows
forks of Chocolate Doom to define their own protocols while retaining
the ability to connect to Chocolate Doom servers (and vice versa, for
Chocolate Doom clients to connect to forked servers).

When a client connects to a server, we send a list of protocols that are
supported. The server compares it to its own list of protocols and
selects a common protocol that will be used. The protocol is then sent
back to the client in a reply packet.

In order to do this, we must change the way that clients connect to
servers. Previously the client sent a SYN packet, and the server would
respond with an ACK packet. Instead, get rid of ACK packets and send a
SYN packet back instead. This also simplifies the state machine when
performing a connect. Since this is a fundamental change in the
protocol, change the magic number sent when connecting, but continue to
recognize the old magic number and send a reject message to old clients.