setup: Tweak server list interface.
When finding a server to join it helps to know if other players are
already connected. So include a player count, but because of the limited
screen space, only if there already are players. Also fix the
capitalization of the word "Internet".

This fixes #907. Thanks to @JNechaevsky for the report.