strife: Move D_IdentifyVersion() earlier in startup.
D_IdentifyVersion() as well as identifying the game version also
loads voices.wad. In order to support WADs which replace lumps from
voices.wad, this needs to be loaded before any PWADs are loaded; this
is also consistent with vanilla Strife behavior. The voices.wad file
was also being loaded after initialization of the WAD file hash table,
which was less efficient.

This fixes #906. Thanks @Catoptromancy for reporting.