Print a more useful error about which savegame failed to load
Amends the prior commit, also fixes the formatting for the I_Error()
function call to represent the preferred format outlined in HACKING.

Hexen now has two locations that can fail out with an error.  One in
CopySaveSlot that should fail when a hex?.hxs file does not exist, as
specified via the -loadgame parameter.  A second in the SV_OpenRead
function that should probably never get triggered after the first; it
would mean the hex6.hxs file wasn't created or somehow deleted in the
middle of the function.  (Chocolate Hexen race condition exploits,
anyone?)