net: Move server reject message into error message.
Since we surface error messages in dialog boxes, it's important that
the full context of why a connection attempt was rejected is shown to
the user, otherwise it just appears as "failed to connect" with no
indication of why or how to fix the problem.

Thanks to AgitationSkeleton and GuyNamedErick for help identifying this
problem.