net: Set keepalive_recv_time on connection setup.
If the keepalive last-receive time is not initialized when the
connection is established, there is a risk that the connection will
immediately time out. This was causing problems for some players
where clients were successfully connecting to servers only to find
themselves stuck in the pre-game "lobby" area with no information
about other players.

This fixes #1046. Thanks to @bradc6 for the bug report.