video: Add -borderless command line parameter.
When playing in three window mode it's potentially helpful to have a
borderless mode so that the user can line up three windows next to
each other on a single desktop. However, without the windows borders
it's difficult/impossible to move windows around, so I'm not sure yet
how useful this actually is. It probably depends on the OS.

Thanks to joshthenesnerd for the suggestion.