glob: let I_NextGlob() return NULL if (glob == NULL)
If a non-existent directory is passed over to I_StartGlob(), e.g. when
moving config files between computers, it will return NULL. If this
return value is then fed into I_NextGlob(), it will trigger a
segmentation fault. Add a check to the latter if (glob == NULL) and
let it return NULL itself in this case.