Refactor mouse speed box.
There was a lot of conditional code here to handle the two different
modes in which it can operate, and these are best served as just two
different functions. Get rid of box_x/box_y variables and replace with
constants.

As part of this I also got rid of the yellow scale indicators in the
non-accelerating case and just made it scale according to the largest
seen mouse speed. They didn't really add anything that I could see,
since the thermometer is really only there to aid calibration in the
case where acceleration is used.