setup: Rearrange mouse config dialog.
This previously showed the fire/use/move forward controls, which are
the original controls provided by Doom's setup.exe.

However, the reality is that nowadays most people play with the WASD&mouse
arrangement (or something similar) and are unlikely to be using the mouse
to move forwards. So it makes more sense if we reserve the main mouse
dialog for "action" buttons (fire/use) and move the "move forward" button
into the more buttons dialog. This also makes the more buttons dialog match
the arrangements for keyboard and gamepad controls.

As part of this, also tweak the mouse button minimum width so that it has a
consistent size that can describe high-numbered mouse buttons (eg.
"BUTTON #10").